refactor(scan): extract `saveSheetImages`
We will need to use this with the NH interpreter, and this lets us reuse the logic for VX and NH ballots.
